Move the sheetViewModel into the ComposeFormDataCollectionFragment.kt
